requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.9
ultralytics>=8.0.0
numpy>=1.24.0
pyyaml>=6.0
supervision>=0.18.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0



pyaudio>=0.2.13
faster-whisper>=0.10.0
piper-tts>=1.2.0
openwakeword>=0.6.0


llama-cpp-python>=0.2.0

bleak>=0.21.0


